+++ backend/src/main/java/com/mikhaile/nostobackend/CurrencyRate.java
@@ -1,7 +1,9 @@
 package com.mikhaile.nostobackend;
 
+import java.math.BigDecimal;
+
 public record CurrencyRate(
     String baseCurrency,
     String quoteCurrency,
-    Float rate) {
+    BigDecimal rate) {
 }
