+++ backend/src/main/java/com/mikhaile/nostobackend/CurrencyRateProviderSwop.java
@@ -5,6 +5,7 @@ import io.vertx.core.Vertx;
 import io.vertx.ext.web.client.WebClient;
 import io.vertx.ext.web.client.WebClientOptions;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.Date;
 
@@ -22,6 +23,7 @@ public class CurrencyRateProviderSwop implements CurrencyRateProvider {
 
     @Override
     public Future<CurrencyRate[]> getAllRates() {
+        // for production, it should also implement retry policy
         return webClient
             .get("/rest/rates")
             .addQueryParam("api-key", apiKey)
@@ -41,7 +43,7 @@ public class CurrencyRateProviderSwop implements CurrencyRateProvider {
     /**
      * API's response JSON mapping
      */
-    private record SwopCurrencyRate(String base_currency, String quote_currency, Float quote, Date date) {
+    private record SwopCurrencyRate(String base_currency, String quote_currency, BigDecimal quote, Date date) {
         public CurrencyRate toCurrencyRate() {
             return new CurrencyRate(base_currency, quote_currency, quote);
         }
